:white_check_mark: Remove `log` test dependency on `log_fmt`
Problem:
- Some `log` tests depend on `log_fmt` and/or `log_binary`. This dependency goes
  the wrong way.

Solution:
- Move some tests from `log` to `log_fmt` and/or `log_binary`.
- Break the remaining `log` test dependency on `log_fmt`.

Note:
- `log_fmt` is still used in `msg` and `flow` tests; it's a useful testing tool.
  And this dependency is OK.
