+++ test/log_binary/CMakeLists.txt
@@ -1,4 +1,11 @@
-add_tests(FILES encoder mipi_logger LIBRARIES cib_log_binary)
+add_tests(
+    FILES
+    encoder
+    level
+    mipi_logger
+    unit
+    LIBRARIES
+    cib_log_binary)
 
 add_library(catalog1_lib STATIC catalog1_lib.cpp)
 add_library(catalog2_lib OBJECT catalog2a_lib.cpp catalog2b_lib.cpp)
