+++ test/log_fmt/CMakeLists.txt
@@ -1,1 +1,1 @@
-add_tests(FILES fmt_logger LIBRARIES cib_log_fmt)
+add_tests(FILES fmt_logger level LIBRARIES cib_log_fmt)
