+++ usage_test/main_log.cpp
@@ -3,6 +3,7 @@
 #include <log/level.hpp>
 #include <log/log.hpp>
 #include <log/module.hpp>
+#include <log/unit.hpp>
 
 #if __STDC_HOSTED__ == 0
 extern "C" auto main() -> int;
